+++ protocols/relay/CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.9.2 - unreleased
+# 0.10.0 [unreleased]
+
+- Update to `libp2p-core` `v0.34.0`.
+
+- Update to `libp2p-swarm` `v0.37.0`.
 
 - Do not duplicate the p2p/xxx component with the relay PeerId when a client requests a reservation. See [PR 2701].