+++ transports/uds/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.33.0 [unreleased]
+
+- Update dependencies.
+- Update to `libp2p-core` `v0.34.0`.
+
 # 0.32.0 [2022-01-27]
 
 - Update to `libp2p-core` `v0.32.0`.